# Conflicts:
#	FP7/FP7Data/fp7orgs.csv

+++ H2020/H2020_Code/edgeStacker.py
@@ -1,5 +1,7 @@
 # Powered by Python 2.7
 
+# RUN THIS FROM THE MAIN
+
 # To cancel the modifications performed by the script
 # on the current graph, click on the undo button.
 
@@ -62,7 +64,7 @@ def findEdge(node1, node2, graph1, directed = False, create = True):
 
 	
 def main(graph): 
-	projectsTogether =  graph.getIntegerProperty("projectsTogether")
+	projectsTogether =  graph.getDoubleProperty("projectsTogether")
 	moneyTogether = graph.getDoubleProperty('moneyTogether')
 	viewLayout =  graph.getLayoutProperty("viewLayout")
 	ShortName =  graph.getStringProperty("ShortName")
@@ -114,6 +116,7 @@ def main(graph):
 	
 	# add all nodes in graph1 to stacked
 	for n in nonStacked.getNodes():
+		
 		stacked.addNode(n)
 		
 	# you go over all edges in graph1 and add only one edge to graph2
